spdrs.com had at least 1 corrupt file. getHoldings.SPDR now checks file and returns NULL if it is corrupt
git-svn-id: svn://scm.r-forge.r-project.org/svnroot/twsinstrument/pkg/qmao@375 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ R/getHoldings.SPDR.R
@@ -85,7 +85,11 @@ getHoldings.SPDR <- function(Symbols, env=.GlobalEnv, auto.assign=TRUE) {
         #stop(paste("Error downloading holdings of SPDR ETF", sQuote(symbol)))
         return(NULL)
     }
-    fr <- read.xls(tmp, skip=3, stringsAsFactors=FALSE)
+    fr <- try(read.xls(tmp, skip=3, stringsAsFactors=FALSE))
+    if (inherits(fr, "try-error")) {
+      warning(paste("There is a problem with holdings file for", sQuote(symbol)))
+      return(NULL)
+    }
     unlink(tmp)
     if (length(colnames(fr)) == 1L) {
         #return(NULL) # HTTP.404..Page.Not.Found
